+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Updated auto-closing for block comments
 - Updated formatter to skip adding newlines to closing curly braces when a semicolon is present after it, and to ignore extra newlines after the opening curly brace.
 - Line comment formatter now ignores URLs
+- Formatter now adds spaces after commas for cleaner look.
 
 ## [0.3.6] - 2025-01-15